requirements.txt
fastapi
uvicorn
sqlmodel
sqlglot
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
aiosqlite
httpx
pytest
pytest-asyncio
urllib3<2.0.0
boto3
pandas
pyarrow
